requirements.txt
anyascii==0.3.2
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
defusedxml==0.7.1
Django==5.1.7
django-filter==25.1
django-modelcluster==6.4
django-permissionedforms==0.1
django-stubs-ext==5.1.3
django-taggit==6.1.0
django-tasks==0.6.1
django-treebeard==4.7.1
djangorestframework==3.15.2
draftjs_exporter==5.1.0
et_xmlfile==2.0.0
filetype==1.2.0
idna==3.10
laces==0.1.2
openpyxl==3.1.5
pillow==11.1.0
pillow_heif==0.22.0
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
telepath==0.3.1
typing_extensions==4.12.2
urllib3==2.3.0
wagtail==6.4.1
Willow==1.9.0
